extensions/carry-forward/pyproject.toml
[project]
name = "carry_forward"
version = "0.0.1"
description = "Canvas plugin to carry forward commands from previous notes"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"

[tool.coverage.run]
source = ["carry_forward"]
omit = ["tests/*", "*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
omit = ["tests/*", "*/tests/*"]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]
